Temperature fix (Firmware > 5.0)
The temperature of the sensor for firmware versions greater than 5 has received a fix.

This changes have only been tested on the MHZ19C.

Please note:  older versions no longer offer a float argument for the getTemperature() function. Instead,  the value returned is in accordance with the data sheet. This is due to the float version being somewhat unreliable.

As a result, temperature values received in older library versions will see an increase by 2C as the offset used in the datasheet differed.

The firmware version (the major version) is obtained on initiation and used to determine which temperature code to run.

Please report any issues as there are many firmware versions and I have only two sensors.